homework-4: refactored components, added category filtering
