+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { CartProvider } from "./CartContext";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import Header from "./components/Header/Header";
+import Footer from "./components/Footer/Footer";
 import MenuPage from "./pages/MenuPage";
 import "./App.css";
