+++ src/components/Header/Header.jsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { CartContext } from "../CartContext";
+import { CartContext } from "../../CartContext";
 import "./Header.css";
 
 function Header() {
