+++ src/pages/MenuPage.jsx
@@ -1,14 +1,14 @@
 import { useState, useEffect } from "react";
-import MenuItem from "../components/MenuItem";
+import MenuItem from "../components/MenuItem/MenuItem";
 import Tooltip from "../components/Tooltip";
 import "./MenuPage.css";
 
 function MenuPage() {
-
   const itemsPerClick = 6;
 
   const [meals, setMeals] = useState([]);
   const [visibleCount, setVisibleCount] = useState(itemsPerClick);
+  const [selectedCategory, setSelectedCategory] = useState(null); 
 
   useEffect(() => {
     fetch("https://65de35f3dccfcd562f5691bb.mockapi.io/api/v1/meals")
@@ -21,6 +21,10 @@ function MenuPage() {
     setVisibleCount((prev) => prev + itemsPerClick);
   };
 
+  const filteredMeals = selectedCategory
+    ? meals.filter((meal) => meal.category === selectedCategory)
+    : meals;
+
   if (meals.length === 0) {
     return <div>Loading...</div>;
   }
@@ -30,26 +34,43 @@ function MenuPage() {
       <div className="menu-container">
         <div className="menu-inner">
           <h2 className="menu-title">Browse our menu</h2>
+
           <div className="menu-subtitle">
-  Use our menu to place an order online, or{" "}
-  <Tooltip tooltipText="+370-12345678">
-    <span className="menu-contact-phone">phone</span>
-  </Tooltip>{" "}
-  our store <br /> to place a pickup order. Fast and fresh food.
-</div>
+            Use our menu to place an order online, or{" "}
+            <Tooltip tooltipText="+370-12345678">
+              <span className="menu-contact-phone">phone</span>
+            </Tooltip>{" "}
+            our store <br /> to place a pickup order. Fast and fresh food.
+          </div>
+
           <div className="menu-buttons">
-            <button className="menu-btn active">Desert</button>
-            <button className="menu-btn">Dinner</button>
-            <button className="menu-btn">Breakfast</button>
+            <button
+              className={`menu-btn ${selectedCategory === "Dessert" ? "active" : ""}`}
+              onClick={() => setSelectedCategory("Dessert")}
+            >
+              Dessert
+            </button>
+            <button
+              className={`menu-btn ${selectedCategory === "Dinner" ? "active" : ""}`}
+              onClick={() => setSelectedCategory("Dinner")}
+            >
+              Dinner
+            </button>
+            <button
+              className={`menu-btn ${selectedCategory === "Breakfast" ? "active" : ""}`}
+              onClick={() => setSelectedCategory("Breakfast")}
+            >
+              Breakfast
+            </button>
           </div>
 
           <div className="menu-items">
-            {meals.slice(0, visibleCount).map((item) => (
+            {filteredMeals.slice(0, visibleCount).map((item) => (
               <MenuItem key={item.id} item={item} />
             ))}
           </div>
 
-          {visibleCount < meals.length && (
+          {visibleCount < filteredMeals.length && (
             <div className="menu-see-more">
               <button className="menu-btn-see-more" onClick={handleSeeMore}>
                 See more
